+++ common/include/common/rfproto.h
@@ -174,6 +174,8 @@ enum RF_GameType
     RF_GT_ESC = 0x07,
     RF_GT_BM = 0x08,
     RF_GT_TBM = 0x09,
+    // Sentinel: must always be the last entry.
+    RF_GT_UNK
 };
 
 enum RF_ServerFlags